Use cudnn handle per device in cudnn layer norm executor
following what the cudnn sdpa ex does
