+++ thunder/executors/cudnn_sdpa.py
@@ -43,33 +43,9 @@ class CudnnexLRUCache(OrderedDict):
 
 
 _cudnnex_cache = CudnnexLRUCache(maxlen=1024)
-# Mapping from device to cudnn handles
-_device_to_cudnn_handle = {}
 
-
-# This function creates a new handle for the device that cudnn should
-# run its kernels on. As the suggested approach by cudnn is to make a few handles
-# as possible, this function caches these per-device handles.
-def _get_cudnn_handle(query_device):
-    handle = _device_to_cudnn_handle.get(query_device, None)
-    if handle is None:
-        with torch.cuda.device(query_device):
-            handle = cudnn.create_handle()
-            _device_to_cudnn_handle[query_device] = handle
-
-    # Make sure the user stream is set on the handle
-    # Fetch the current user stream and pass the data pointer to set_stream API
-    cudnn.set_stream(handle=handle, stream=torch.cuda.current_stream(device=query_device).cuda_stream)
-
-    return handle
-
-    handle = _device_to_cudnn_handle.get(query_device, None)
-    if handle is None:
-        with torch.cuda.device(query_device):
-            handle = cudnn.create_handle()
-            _device_to_cudnn_handle[query_device] = handle
-    cudnn.set_stream(handle=handle, stream=torch.cuda.current_stream(device=query_device).cuda_stream)
-    return handle
+# Import shared cuDNN handle management
+from thunder.executors.cudnnex import _get_cudnn_handle
 
 
 def _make_cudnn_sdpa_forward_graph(
